Extract class Mesh to Mesh.h and Mesh.cpp
