+++ Agate/src/Agate/Rendering/Mesh.h
@@ -1,0 +1,66 @@
+/**
+ * @brief Include file for the Mesh class
+ * 
+ * Has render data for OpenGL and the ability to invoke OpenGL
+ * to draw itself
+ */
+
+#ifndef AGATE_MESH_H
+#define AGATE_MESH_H
+
+#include "OpenGl/Shader.h"
+#include "OpenGl/Texture.h"
+#include "OpenGl/VertexArray.h"
+#include <glm/glm.hpp>
+#include <memory>
+#include <vector>
+
+namespace Agate {
+
+/**
+ * @brief A vertex with a position, normal, texture coordinate,
+ *        tanget, and bitangent
+ */
+struct Vertex {
+    glm::vec3 Position;
+    glm::vec3 Normal;
+    glm::vec2 TexCoords;
+    glm::vec3 Tangent;
+    glm::vec3 Bitangent;
+};
+
+class Mesh {
+public:
+
+    std::vector<Vertex> vertices;
+    std::vector<unsigned int> indices;
+    std::vector<Texture> textures;
+
+    /**
+     * @brief Constructor from mesh data
+     * 
+     * @param vertices Vertices that make up the mesh
+     * @param indices Relative indices of the mesh vertices to map to faces
+     * @param textures Textures required by this mesh
+     */
+    Mesh(std::vector<Vertex> vertices, std::vector<unsigned int> indices, std::vector<Texture> textures);
+
+    /**
+     * @brief Draws this mesh
+     * 
+     * @param shader Shader to bind and use for shading the mesh
+     */
+    void Draw(Shader &shader);
+
+    virtual ~Mesh();
+
+private:
+    //  render data
+    std::shared_ptr<VertexArray> VA;
+
+    void setupMesh();
+};
+
+}
+
+#endif // AGATE_MESH_H
